Completed adding a default logo to the title bar
