+++ src/app/molecules/productImage/styles.css
@@ -9,3 +9,4 @@
     height: 310px;
     width: 227px;
 }
+
