--- src/app/molecules/productType/index.tsx
@@ -1,1 +1,0 @@
-//Here we will pass all the props down to each component: Logo, image, tips, title
