+++ src/app/molecules/tips/index.tsx
@@ -3,9 +3,11 @@ import './styles.css'
 
 export const TipsMolecule: React.FC = () => {
 
+//Based on product chosen change the logo and the text rendered
+
     return (
         <div>
-
+            tips
         </div>
     )
 }
