+++ src/app/molecules/titleBar/index.tsx
@@ -1,13 +1,28 @@
 import './styles'
 import { SettingsButton } from '../../atoms/buttons'
+import JPIVetLogo from '../../../../public/assets/image/logos/JPIVetLogo.png'
 
 const TitleBarMolecule = () => {
+
+/*Based on the context, 
+- Change the JPI Product to the Product chosen
+- Change the icon to the product chosen
+*/
+
+
+
+
     return (
         <div className="titleBarMoleContainer">
-            <div className="Title">
-                JPI Product | Operation Tips
+            <div className="titleContent">
+                <div className="titleBarLogo">
+                    <img src={JPIVetLogo} alt="JPI Vet Logo"/>
+                </div>
+                <div>
+                DirectVet | Operation Tips
+                </div>
             </div>
-            <div className="Settings">
+            <div className="SettingsButton">
                 <SettingsButton />
             </div>
         </div>
