+++ src/app/organisms/tips/index.tsx
@@ -1,0 +1,13 @@
+import ProductImage from "../../molecules/productImage"
+import { TipsMolecule } from "../../molecules/tips"
+
+const TipsOrganism = () => {
+    return (
+            <div className="pageItems">
+                <ProductImage />
+                <TipsMolecule />
+            </div>
+    )
+}
+
+export default TipsOrganism
