--- src/app/organisms/tips/tips.tsx
@@ -1,12 +1,0 @@
-import { TipsMolecule } from "../../molecules/tips"
-
-const Tips = () => {
-  return (
-    <div>
-        <TipsMolecule/>
-        Tips 
-        </div>
-  )
-}
-
-export default Tips
